serverless-vllm-manager/requirements.txt
mcp
google-cloud-logging
google-cloud-monitoring
vllm
pydantic>=2.0
